+++ cli/config.lua
@@ -3,14 +3,23 @@
 -- Sledmine
 -- Create and provide environment stuff to use over the code
 ------------------------------------------------------------------------------
-local environment = {}
-local lfs = require "lfs"
+local config = {}
 local glue = require "glue"
 local json = require "cjson"
 local registry = require "registry"
 
 -- Paths table instance
 local paths
+local cfgPath = gpath(exedir(), "/mercuryconf.json")
+
+---@class configCLI
+local cfg = {
+    game = {
+        path = nil,
+        data = {path = nil}
+        -- , maps = {path = nil}, mods = {path = nil}
+    }
+}
 
 -- Windows required registry keys
 local registryEntries = {
@@ -34,14 +43,6 @@ local function getGamePath()
             gamePath = query.values["EXE Path"]["value"]
         end
     end
-    if (not gamePath) then
-        cprint("Error, Halo Custom Edition path was not found on the system.")
-        cprint("Force game path by setting \"HALO_CE_PATH\" as an environment variable.\n")
-        cprint("Example:")
-        cprint([[On Linux: export HALO_CE_PATH="/home/117/.wine/c/Halo Custom Edition"]])
-        cprint([[On Windows: set HALO_CE_PATH=D:\Games\Halo Custom Edition]])
-        os.exit(1)
-    end
     return gamePath
 end
 
@@ -54,58 +55,54 @@ local function getMyGamesHaloCEPath()
             documentsPath = query.values["Personal"]["value"] .. "\\My Games\\Halo CE"
         end
     end
-    if (not documentsPath) then
-        cprint("Error, at trying to get \"My Games\" path from the system.")
-        cprint(
-            "Force game path by setting \"MY_GAMES_PATH\" or \"HALO_CE_DATA_PATH\" as an environment variable.\n")
-        cprint("Example:")
-        cprint([[On Linux: export MY_GAMES_PATH="/home/117/Documents/My Games/Halo CE"]])
-        cprint([[On Windows: set MY_GAMES_PATH=D:\Users\117\Documents\My Games\Halo CE]])
-        os.exit(1)
-    end
     return documentsPath
 end
 
 --- Setup environment to work, environment variables, configuration folder, etc
-function environment.paths()
+function config.paths()
     -- local sourceFolder = lfs.currentdir()
     -- local appData = os.getenv("APPDATA")
 
     -- Singleton like method, return gathered paths instead of getting them every invocation
-    if (not paths) then
-        local gamePath = gpath(getGamePath())
-        local myGamesHaloCEPath = gpath(getMyGamesHaloCEPath())
-        local mercuryTemp = gpath((os.getenv("TEMP") or "/tmp") .. "/mercury")
-        local mercuryDownloads = gpath(mercuryTemp, "/downloads")
-        local mercuryUnpacked = gpath(mercuryTemp, "/unpacked")
-        local mercuryOldIndex = gpath(gamePath, "/mercury/installed/packages.json")
-        local mercuryIndex = gpath(gamePath, "/mercury.json")
-        local luaScriptsGlobal = gpath(myGamesHaloCEPath, "/chimera/lua/scripts/global")
-        local luaScriptsMap = gpath(myGamesHaloCEPath, "/chimera/lua/scripts/map")
-        local luaScriptsSAPP = gpath(myGamesHaloCEPath, "/sapp/lua")
-        local luaDataGlobal = gpath(myGamesHaloCEPath, "/chimera/lua/data/global")
-        local luaDataMap = gpath(myGamesHaloCEPath, "/chimera/lua/data/map")
-        local gameMaps = gpath(gamePath, "/maps")
-        local gameDLLMods = gpath(gamePath, "/mods")
+    if not paths then
+        local game = cfg.game.path or gpath(getGamePath())
+        local data = cfg.game.data.path or gpath(getMyGamesHaloCEPath())
 
-        createFolder(mercuryDownloads)
-        createFolder(mercuryUnpacked)
-        paths = {
-            gamePath = gamePath,
-            myGamesPath = myGamesHaloCEPath,
-            mercuryTemp = mercuryTemp,
-            mercuryDownloads = mercuryDownloads,
-            mercuryUnpacked = mercuryUnpacked,
-            mercuryOldIndex = mercuryOldIndex,
-            mercuryIndex = mercuryIndex,
-            luaScriptsGlobal = luaScriptsGlobal,
-            luaScriptsMap = luaScriptsMap,
-            luaScriptsSAPP = luaScriptsSAPP,
-            luaDataGlobal = luaDataGlobal,
-            luaDataMap = luaDataMap,
-            gameMaps = gameMaps,
-            gameDLLMods = gameDLLMods
-        }
+        if game and data then
+            local maps = gpath(game, "/maps")
+            local mods = gpath(game, "/mods")
+
+            local mercuryTemp = gpath((getenv "TEMP" or "/tmp") .. "/mercury")
+            local mercuryDownloads = gpath(mercuryTemp, "/downloads")
+            local mercuryUnpacked = gpath(mercuryTemp, "/unpacked")
+            local mercuryIndex = gpath(game, "/mercury.json")
+            local mercuryOldIndex = gpath(game, "/mercury/installed/packages.json")
+
+            local luaScriptsGlobal = gpath(data, "/chimera/lua/scripts/global")
+            local luaScriptsMap = gpath(data, "/chimera/lua/scripts/map")
+            local luaScriptsSAPP = gpath(data, "/sapp/lua")
+            local luaDataGlobal = gpath(data, "/chimera/lua/data/global")
+            local luaDataMap = gpath(data, "/chimera/lua/data/map")
+
+            createFolder(mercuryDownloads)
+            createFolder(mercuryUnpacked)
+            paths = {
+                gamePath = game,
+                myGamesPath = data,
+                mercuryTemp = mercuryTemp,
+                mercuryDownloads = mercuryDownloads,
+                mercuryUnpacked = mercuryUnpacked,
+                mercuryOldIndex = mercuryOldIndex,
+                mercuryIndex = mercuryIndex,
+                luaScriptsGlobal = luaScriptsGlobal,
+                luaScriptsMap = luaScriptsMap,
+                luaScriptsSAPP = luaScriptsSAPP,
+                luaDataGlobal = luaDataGlobal,
+                luaDataMap = luaDataMap,
+                gameMaps = maps,
+                gameDLLMods = mods
+            }
+        end
     end
     return paths
 end
@@ -113,11 +110,11 @@ end
 --- Get mercury local installed packages
 ---@param newPackages? packageMercury[]
 ---@return packageMercury[]? packages
-function environment.packages(newPackages)
+function config.packages(newPackages)
     if (not newPackages) then
         if (exists(paths.mercuryIndex)) then
             local installedPackages = json.decode(readFile(paths.mercuryIndex))
-            if (installedPackages and #glue.keys(installedPackages) > 0) then
+            if (installedPackages and #table.keys(installedPackages) > 0) then
                 return installedPackages
             end
         end
@@ -130,7 +127,7 @@ function environment.packages(newPackages)
 end
 
 --- Clean temp data, temp folders, trash files...
-function environment.clean()
+function config.clean()
     if (not IsDebugModeEnabled) then
         dprint("Cleaning " .. paths.mercuryTemp .. "...")
         delete(paths.mercuryTemp, true)
@@ -140,7 +137,7 @@ function environment.clean()
 end
 
 --- Migrate deprecated or old files and paths
-function environment.migrate()
+function config.migrate()
     if (exists(paths.mercuryOldIndex)) then
         cprint("Warning, migrating old packages index path to new index path!")
         move(paths.mercuryOldIndex, paths.mercuryIndex)
@@ -148,4 +145,66 @@ function environment.migrate()
     end
 end
 
-return environment
+function config.load()
+    local currentCfg
+    if exists(cfgPath) then
+        currentCfg = json.decode(readFile(cfgPath))
+        if currentCfg then
+            cfg = table.merge(cfg, currentCfg) --[[@as configCLI]]
+        end
+    end
+    config.paths()
+    return currentCfg
+end
+
+--- Get CLI configuration
+---@param key string?
+---@return configCLI
+function config.get(key)
+    if key then
+        if key:includes "." then
+            local keys = key:split "."
+            local value = cfg
+            for _, k in ipairs(keys) do
+                value = value[k]
+            end
+            return value
+        else
+            return cfg[key]
+        end
+    end
+    return cfg
+end
+
+--- Set CLI configuration
+---@param key string
+---@param value any
+---@return boolean
+function config.set(key, value)
+    if key then
+        if value == "" then
+            value = nil
+        end
+        if key:includes "." then
+            local keys = key:split "."
+            local lastKey = keys[#keys]
+            local currentCfg = cfg
+            for _, k in ipairs(keys) do
+                if k == lastKey then
+                    currentCfg[k] = value
+                else
+                    currentCfg = currentCfg[k]
+                end
+            end
+        else
+            cfg[key] = value
+        end
+        if writeFile(cfgPath, json.encode(cfg)) then
+            print(cfgPath)
+            return true
+        end
+    end
+    return false
+end
+
+return config